+++ kart/working_copy/gpkg.py
@@ -10,7 +10,6 @@ from osgeo import gdal
 import sqlalchemy as sa
 from sqlalchemy.dialects.sqlite.base import SQLiteIdentifierPreparer
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.types import UserDefinedType
 
 
 from . import WorkingCopyStatus
@@ -18,7 +17,6 @@ from .base import BaseWorkingCopy
 from .table_defs import GpkgTables, GpkgKartTables
 from kart import crs_util
 from kart.exceptions import InvalidOperation
-from kart.geometry import normalise_gpkg_geom
 from kart.schema import Schema
 from kart.sqlalchemy import text_with_inlined_params
 from kart.sqlalchemy.adapter.gpkg import KartAdapter_GPKG
@@ -82,16 +80,6 @@ class WorkingCopy_GPKG(BaseWorkingCopy):
     def full_path(self):
         """ Return a full absolute path to the working copy """
         return (self.repo.workdir_path / self.path).resolve()
-
-    def _type_def_for_column_schema(self, col, dataset):
-        if col.data_type == "geometry":
-            # This user-defined GeometryType normalises GPKG geometry to the Kart V2 GPKG geometry.
-            return GeometryType
-        elif col.data_type == "boolean":
-            # Read BOOLEANs as bools, not ints.
-            return BooleanType
-        # Don't need to specify type information for other columns at present, since we just pass through the values.
-        return None
 
     @property
     def _tracking_table_requires_cast(self):
@@ -655,33 +643,3 @@ class WorkingCopy_GPKG(BaseWorkingCopy):
                 {"last_change": gpkg_change_time, "table_name": dataset.table_name},
             ).rowcount
         assert rc == 1, f"gpkg_contents update: expected 1Δ, got {rc}"
-
-
-class GeometryType(UserDefinedType):
-    """UserDefinedType so that GPKG geometry is normalised to V2 format."""
-
-    def result_processor(self, dialect, coltype):
-        def process(gpkg_bytes):
-            # Its possible in GPKG to put arbitrary values in columns, regardless of type.
-            # We don't try to convert them here - we let the commit validation step report this as an error.
-            if not isinstance(gpkg_bytes, bytes):
-                return gpkg_bytes
-            # We normalise geometries to avoid spurious diffs - diffs where nothing
-            # of any consequence has changed (eg, only endianness has changed).
-            # This includes setting the SRID to zero for each geometry so that we don't store a separate SRID per geometry,
-            # but only one per column at most.
-            return normalise_gpkg_geom(gpkg_bytes)
-
-        return process
-
-
-class BooleanType(UserDefinedType):
-    """UserDefinedType so that BOOLEANs are read as bools and not ints."""
-
-    def result_processor(self, dialect, coltype):
-        def process(gpkg_int):
-            # Its possible in GPKG to put arbitrary values in columns, regardless of type.
-            # We don't try to convert them here - we let the commit validation step report this as an error.
-            return bool(gpkg_int) if gpkg_int in (0, 1) else gpkg_int
-
-        return process